+++ R/bd_retrieve.R
@@ -1,0 +1,3 @@
+bd_retrieve <- function(id) {
+  bd_retrieve_data(searchTerm = as.character(id), searchType = "bacdive_id")
+}