+++ R/bd_retrieve_by_culture.R
@@ -1,0 +1,3 @@
+bd_retrieve_by_culture <- function(collection_no) {
+  bd_retrieve_data(searchTerm = collection_no, searchType = "culturecollectionno")
+}