+++ R/bd_retrieve_by_sequence.R
@@ -1,0 +1,3 @@
+bd_retrieve_by_sequence <- function(accession) {
+  bd_retrieve_data(searchTerm = accession, searchType = "sequence")
+}