+++ R/bd_retrieve_by_taxon.R
@@ -1,0 +1,3 @@
+bd_retrieve_by_taxon <- function(name) {
+  bd_retrieve_data(searchTerm = name, searchType = "taxon")
+}