+++ R/bd_retrieve_taxon.R
@@ -4,9 +4,9 @@
 #' @export
 #' @examples
 #' \donttest{
-#' datasets_Bh <- bd_retrieve_by_taxon(name = "Bacillus halotolerans")
-#' datasets_Bss <- bd_retrieve_by_taxon(name = "Bacillus subtilis subtilis")
+#' datasets_Bh <- bd_retrieve_taxon(name = "Bacillus halotolerans")
+#' datasets_Bss <- bd_retrieve_taxon(name = "Bacillus subtilis subtilis")
 #' }
-bd_retrieve_by_taxon <- function(name) {
+bd_retrieve_taxon <- function(name) {
   bd_retrieve_data(searchTerm = name, searchType = "taxon")
 }