+++ NEWS.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 
-## Bacdive 0.8.0
+## BacDiveR 0.8.0
 
 ### Added